Final_Project/requirements.txt
streamlit
python-dotenv
langchain
langchain-google-genai
langchain-huggingface
langchain-community
langchain-chroma
langchain-text-splitters
transformers
torch
pypdf  
chromadb
pillow
PyMuPDF  # For comprehensive PDF extraction